+++ firmware/include/modes/ClockMode.h
@@ -39,9 +39,9 @@ private:
     bool pending{false};
     bool strikethrough{false};
 
-    int hour{24};
-    int minute{60};
-    int second{60};
+    uint8_t hour{24U};
+    uint8_t minute{60U};
+    uint8_t second{60U};
 
     void drawTicker(uint8_t brightness) const;
     void setFont(std::string_view _fontName);
