+++ firmware/src/modes/BinaryClockMode.cpp
@@ -20,17 +20,17 @@ void BinaryClockMode::handle()
         }
         if (second != local.tm_sec || pending)
         {
-            second = local.tm_sec;
+            second = static_cast<uint8_t>(local.tm_sec);
             draw((GRID_ROWS / 2U) + 3U, second);
         }
         if (minute != local.tm_min || pending)
         {
-            minute = local.tm_min;
+            minute = static_cast<uint8_t>(local.tm_min);
             draw((GRID_ROWS / 2U) - 2U, minute);
         }
         if (hour != local.tm_hour || pending)
         {
-            hour = local.tm_hour;
+            hour = static_cast<uint8_t>(local.tm_hour);
             draw((GRID_ROWS / 2U) - 7U, hour);
             pending = false;
         }
